backend/pyproject.toml
[project]
name = "toy-dt-v2"
version = "0.1.0"
description = "TWIN4DEM as a Django web app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "channels-redis>=4.3.0",
    "channels[daphne]>=4.3.2",
    "daphne>=4.2.1",
    "django>=5.2.8",
    "django-allauth[socialaccount]>=65.13.1",
    "django-sass-processor>=1.4.2",
    "django-vite>=3.1.0",
    "djangorestframework>=3.16.1",
    "jwt>=1.4.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
]

[tool.uv]
package = true

[tool.uv.build-backend]
namespace = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = [".", "src"]

[dependency-groups]
dev = [
    "anyio>=4.12.0",
    "black>=25.12.0",
    "django-compressor>=4.6.0",
    "django-debug-toolbar>=6.1.0",
    "libsass>=0.23.0",
    "pytest>=9.0.1",
    "pytest-async>=0.1.1",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "ruff>=0.14.9",
    "types-channels>=4.3.0.20250822",
]
prod = []
